+++ catalystwan/endpoints/configuration/policy/list/prefix.py
@@ -3,19 +3,12 @@
 # mypy: disable-error-code="empty-body"
 from uuid import UUID
 
+from catalystwan.api.templates.models.cisco_vpn_model import PrefixList
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import PrefixList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.prefix import PrefixListEditPayload, PrefixListInfo
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class PrefixListEditPayload(PrefixList, PolicyListId):
-    pass
-
-
-class PrefixListInfo(PrefixList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyPrefixList(APIEndpoints, PolicyListEndpoints):
