+++ catalystwan/models/policy/__init__.py
@@ -24,6 +24,7 @@ from catalystwan.models.policy.list.mirror import MirrorList
 from catalystwan.models.policy.list.policer import PolicerList
 from catalystwan.models.policy.list.port import PortList
 from catalystwan.models.policy.list.preferred_color_group import PreferredColorGroupList
+from catalystwan.models.policy.list.prefix import PrefixList
 
 from .centralized import CentralizedPolicy, TrafficDataDirection
 from .definition.access_control_list import AclPolicy
@@ -41,7 +42,6 @@ from .definition.traffic_data import TrafficDataPolicy
 from .definition.vpn_membership import VPNMembershipPolicy
 from .definition.zone_based_firewall import ZoneBasedFWPolicy
 from .lists import (
-    PrefixList,
     ProtocolNameList,
     RegionList,
     SiteList,
